+++ simulations/vip-530/ethereum.ts
@@ -5,7 +5,7 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents, setMaxStalePeriod } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip525, { wstETH } from "../../vips/vip-525/bscmainnet";
+import vip530, { wstETH } from "../../vips/vip-530/bscmainnet";
 import ERC20_ABI from "./abi/ERC20.json";
 import RESILIENT_ORACLE_ABI from "./abi/ResilientOracle.json";
 
@@ -15,7 +15,7 @@ const prices = [
   {
     symbol: "wstETH",
     address: wstETH,
-    expectedPrice: parseUnits("2957.471395275213546772", 18),
+    expectedPrice: parseUnits("3211.216687207824390599", 18),
     postVIP: async function (resilientOracle: any) {
       const token = new ethers.Contract("0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2", ERC20_ABI, ethers.provider);
       await setMaxStalePeriod(resilientOracle, token);
@@ -23,7 +23,7 @@ const prices = [
   },
 ];
 
-forking(22816596, async () => {
+forking(22882203, async () => {
   const provider = ethers.provider;
   const resilientOracle = new ethers.Contract(ethereum.RESILIENT_ORACLE, RESILIENT_ORACLE_ABI, provider);
   describe("Pre-VIP behaviour", async () => {
@@ -34,7 +34,7 @@ forking(22816596, async () => {
     }
   });
 
-  testForkedNetworkVipCommands("vip525", await vip525(), {
+  testForkedNetworkVipCommands("vip530", await vip530(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [RESILIENT_ORACLE_ABI], ["TokenConfigAdded"], [1]);
     },
