+++ vips/vip-530/bsctestnet.ts
@@ -5,27 +5,26 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { LzChainId, ProposalType } from "src/types";
 import { makeProposal } from "src/utils";
 
-const { ethereum } = NETWORK_ADDRESSES;
+const { sepolia } = NETWORK_ADDRESSES;
 
-export const PTsUSDe_ORACLE = "0xC407403fa78Bce509821D776b6Be7f91cC04474f";
-export const PTsUSDE_26JUN2025 = "0xDD809435ba6c9d6903730f923038801781cA66ce";
-export const SUSDE_ONEJUMP_CHAINLINK_ORACLE = "0xA67F01322AF8EBa444D788Ee398775b446de51a0";
-export const SUSDE_ONEJUMP_REDSTONE_ORACLE = "0x2B2895104f958E1EC042E6Ba5cbfeCbAD3C5beDb";
-export const SUSDE = "0x211Cc4DD073734dA055fbF44a2b4667d5E5fE5d2";
-export const USDE = "0x5d3a1ff2b6bab83b63cd9ad0787074081a52ef34";
-export const XSOLVBTC_ONEJUMP_REDSTONE_ORACLE = "0xf5534f78Df9b610B19A63956d498d00CFaD8B9D3";
-export const XSOLVBTC = "0x1346b618dC92810EC74163e4c27004c921D446a5";
-export const AnkrBNB_ORACLE = "0x4512e9579734f7B8730f0a05Cd0D92DC33EB2675";
-export const AnkrBNB = "0x52f24a5e03aee338da5fd9df68d2b6fae1178827";
-export const AsBNB_ORACLE = "0x652B90D1d45a7cD5BE82c5Fb61a4A00bA126dde5";
-export const AsBNB = "0x77734e70b6E88b4d82fE632a168EDf6e700912b6";
-export const BNBx_ORACLE = "0xC2E2b6f9CdE2BFA5Ba5fda2Dd113CAcD781bdb31";
-export const BNBx = "0x1bdd3cf7f79cfb8edbb955f20ad99211551ba275";
-export const SlisBNB_ORACLE = "0xDDE6446E66c786afF4cd3D183a908bCDa57DF9c1";
-export const SlisBNB = "0xB0b84D294e0C75A6abe60171b70edEb2EFd14A1B";
+export const PTsUSDE26JUN2025Oracle = "0x85201328baa52061E6648d9b4c285529411Cd33B";
+export const PTsUSDE26JUN2025 = "0x95e58161BA2423c3034658d957F3f5b94DeAbf81";
+export const sUSDe_Chainlink_Oracle = "0x4678BcB5B8eDd9f853725F64d59Ba592F9e41565";
+export const sUSDe_Redstone_Oracle = "0xA5b51bF1625c1F90341c4527AFa5B0865F15ac70";
+export const sUSDe = "0xcFec590e417Abb378cfEfE6296829E35fa25cEbd";
+export const xSolvBTCOracle = "0x33a2BDcBB401a81C590215a6953A9a4B90aD57b9";
+export const xSolvBTC = "0x3ea87323806586A0282b50377e0FEa76070F532B";
+export const ankrBNBOracle = "0x7655d558c3C865913013D82fF4d1e70e97cDf906";
+export const ankrBNB = "0x5269b7558D3d5E113010Ef1cFF0901c367849CC9";
+export const asBNBOracle = "0xb31909f6687Da5bEc559DB7baeed41E14f5Dc17E";
+export const asBNB = "0xc625f060ad25f4A6c2d9eBF30C133dB61B7AF072";
+export const BNBxOracle = "0x068945930785e6816faE855a2A2e8c59BAD380f0";
+export const BNBx = "0x327d6E6FAC0228070884e913263CFF9eFed4a2C8";
+export const slisBNBOracle = "0x6a8154699b6670Ba6166ba59d1c094663E603cA8";
+export const slisBNB = "0xd2aF6A916Bc77764dc63742BC30f71AF4cF423F4";
 
-export const wstETH = "0x7f39C581F595B53c5cb19bD0b3f8dA6c935E2Ca0";
-export const wstETH_ORACLE = "0x6b51Ee3aF70b350AaADc05f418502b330c5Aad7c";
+export const wstETH = "0x9b87ea90fdb55e1a0f17fbeddcf7eb0ac4d50493";
+export const wstETH_ORACLE = "0xA33f06dB4c0DD8E41BA3d2C8917a27e6B9200662";
 
 export const DAYS_30 = 30 * 24 * 60 * 60;
 export const increaseExchangeRateByPercentage = (
@@ -45,41 +44,41 @@ export const getSnapshotGap = (
 };
 
 export const SECONDS_PER_YEAR = 31536000;
-export const sUSDe_Chainlink_InitialExchangeRate = parseUnits("1.179598528113768899", 18);
+export const sUSDe_Chainlink_InitialExchangeRate = parseUnits("1.1", 18);
 export const sUSDe_Chainlink_Timestamp = 1751269105;
-export const sUSDe_Redstone_InitialExchangeRate = parseUnits("1.17966662", 18);
+export const sUSDe_Redstone_InitialExchangeRate = parseUnits("1.1", 18);
 export const sUSDe_Redstone_Timestamp = 1751269105;
 export const sUSDe_GrowthRate = parseUnits("0.2827", 18); // 28.27% per year
 export const sUSDe_SnapshotGap = 236; // 2.36%
-export const PTsUSDE26JUN2025_InitialExchangeRate = parseUnits("0.847697123107543722", 18);
+export const PTsUSDE26JUN2025_InitialExchangeRate = parseUnits("0.85", 18);
 export const PTsUSDE26JUN2025_Timestamp = 1751269105;
 export const PTsUSDE26JUN2025_GrowthRate = SECONDS_PER_YEAR; // 0% per yeat
 export const PTsUSDE26JUN2025_SnapshotGap = 400; // 4.00%
-export const xSolvBTC_InitialExchangeRate = parseUnits("1", 18);
+export const xSolvBTC_InitialExchangeRate = parseUnits("1.", 18);
 export const xSolvBTC_Timestamp = 1751269105;
 export const xSolvBTC_GrowthRate = SECONDS_PER_YEAR; // 0%
 export const xSolvBTC_SnapshotGap = 400; // 4.00%
-export const BNBx_InitialExchangeRate = parseUnits("1.104244614844799120", 18);
+export const BNBx_InitialExchangeRate = parseUnits("1.857886913843483194", 18);
 export const BNBx_Timestamp = 1751269105;
 export const BNBx_GrowthRate = parseUnits("0.0753", 18); // 7.53% per year
 export const BNBx_SnapshotGap = 63; // 0.63%
-export const ankrBNB_InitialExchangeRate = parseUnits("1.097701721231826695", 18);
+export const ankrBNB_InitialExchangeRate = parseUnits("1.080640588742602582", 18);
 export const ankrBNB_Timestamp = 1751269105;
 export const ankrBNB_GrowthRate = parseUnits("0.0612", 18); // 6.12% per year
 export const ankrBNB_SnapshotGap = 51; // 0.51%
-export const slisBNB_InitialExchangeRate = parseUnits("1.030088514324890934", 18);
+export const slisBNB_InitialExchangeRate = parseUnits("4.192864677240396787", 18);
 export const slisBNB_Timestamp = 1751269105;
 export const slisBNB_GrowthRate = parseUnits("0.0412", 18); // 4.12% per year
 export const slisBNB_SnapshotGap = 35; // 0.35%
-export const asBNB_InitialExchangeRate = parseUnits("1.026229121794947588", 18);
+export const asBNB_InitialExchangeRate = parseUnits("1", 18);
 export const asBNB_Timestamp = 1751269105;
 export const asBNB_GrowthRate = parseUnits("0.3049", 18); // 30.49% per year
 export const asBNB_SnapshotGap = 254; // 2.54%
 
-export const vip525 = () => {
+export const vip530 = () => {
   const meta = {
     version: "v2",
-    title: "VIP-525",
+    title: "VIP-530",
     description: "",
     forDescription: "Execute this proposal",
     againstDescription: "Do not execute this proposal",
@@ -89,7 +88,7 @@ export const vip525 = () => {
   return makeProposal(
     [
       {
-        target: SUSDE_ONEJUMP_CHAINLINK_ORACLE,
+        target: sUSDe_Chainlink_Oracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(sUSDe_Chainlink_InitialExchangeRate, BigNumber.from(sUSDe_SnapshotGap)),
@@ -97,17 +96,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: SUSDE_ONEJUMP_CHAINLINK_ORACLE,
+        target: sUSDe_Chainlink_Oracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [sUSDe_GrowthRate, DAYS_30],
       },
       {
-        target: SUSDE_ONEJUMP_CHAINLINK_ORACLE,
+        target: sUSDe_Chainlink_Oracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(sUSDe_Chainlink_InitialExchangeRate, sUSDe_SnapshotGap)],
       },
       {
-        target: SUSDE_ONEJUMP_REDSTONE_ORACLE,
+        target: sUSDe_Redstone_Oracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(sUSDe_Redstone_InitialExchangeRate, BigNumber.from(sUSDe_SnapshotGap)),
@@ -115,17 +114,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: SUSDE_ONEJUMP_REDSTONE_ORACLE,
+        target: sUSDe_Redstone_Oracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [sUSDe_GrowthRate, DAYS_30],
       },
       {
-        target: SUSDE_ONEJUMP_REDSTONE_ORACLE,
+        target: sUSDe_Redstone_Oracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(sUSDe_Redstone_InitialExchangeRate, sUSDe_SnapshotGap)],
       },
       {
-        target: PTsUSDe_ORACLE,
+        target: PTsUSDE26JUN2025Oracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(
@@ -136,17 +135,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: PTsUSDe_ORACLE,
+        target: PTsUSDE26JUN2025Oracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [PTsUSDE26JUN2025_GrowthRate, DAYS_30],
       },
       {
-        target: PTsUSDe_ORACLE,
+        target: PTsUSDE26JUN2025Oracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(PTsUSDE26JUN2025_InitialExchangeRate, PTsUSDE26JUN2025_SnapshotGap)],
       },
       {
-        target: XSOLVBTC_ONEJUMP_REDSTONE_ORACLE,
+        target: xSolvBTCOracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(xSolvBTC_InitialExchangeRate, BigNumber.from(xSolvBTC_SnapshotGap)),
@@ -154,17 +153,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: XSOLVBTC_ONEJUMP_REDSTONE_ORACLE,
+        target: xSolvBTCOracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [xSolvBTC_GrowthRate, DAYS_30],
       },
       {
-        target: XSOLVBTC_ONEJUMP_REDSTONE_ORACLE,
+        target: xSolvBTCOracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(xSolvBTC_InitialExchangeRate, xSolvBTC_SnapshotGap)],
       },
       {
-        target: BNBx_ORACLE,
+        target: BNBxOracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(BNBx_InitialExchangeRate, BigNumber.from(BNBx_SnapshotGap)),
@@ -172,17 +171,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: BNBx_ORACLE,
+        target: BNBxOracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [BNBx_GrowthRate, DAYS_30],
       },
       {
-        target: BNBx_ORACLE,
+        target: BNBxOracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(BNBx_InitialExchangeRate, BNBx_SnapshotGap)],
       },
       {
-        target: AnkrBNB_ORACLE,
+        target: ankrBNBOracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(ankrBNB_InitialExchangeRate, BigNumber.from(ankrBNB_SnapshotGap)),
@@ -190,17 +189,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: AnkrBNB_ORACLE,
+        target: ankrBNBOracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [ankrBNB_GrowthRate, DAYS_30],
       },
       {
-        target: AnkrBNB_ORACLE,
+        target: ankrBNBOracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(ankrBNB_InitialExchangeRate, ankrBNB_SnapshotGap)],
       },
       {
-        target: SlisBNB_ORACLE,
+        target: slisBNBOracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(slisBNB_InitialExchangeRate, BigNumber.from(slisBNB_SnapshotGap)),
@@ -208,17 +207,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: SlisBNB_ORACLE,
+        target: slisBNBOracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [slisBNB_GrowthRate, DAYS_30],
       },
       {
-        target: SlisBNB_ORACLE,
+        target: slisBNBOracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(slisBNB_InitialExchangeRate, slisBNB_SnapshotGap)],
       },
       {
-        target: AsBNB_ORACLE,
+        target: asBNBOracle,
         signature: "setSnapshot(uint256,uint256)",
         params: [
           increaseExchangeRateByPercentage(asBNB_InitialExchangeRate, BigNumber.from(asBNB_SnapshotGap)),
@@ -226,17 +225,17 @@ export const vip525 = () => {
         ],
       },
       {
-        target: AsBNB_ORACLE,
+        target: asBNBOracle,
         signature: "setGrowthRate(uint256,uint256)",
         params: [asBNB_GrowthRate, DAYS_30],
       },
       {
-        target: AsBNB_ORACLE,
+        target: asBNBOracle,
         signature: "setSnapshotGap(uint256)",
         params: [getSnapshotGap(asBNB_InitialExchangeRate, asBNB_SnapshotGap)],
       },
       {
-        target: ethereum.RESILIENT_ORACLE,
+        target: sepolia.RESILIENT_ORACLE,
         signature: "setTokenConfigs((address,address[3],bool[3],bool)[])",
         params: [
           [
@@ -248,7 +247,7 @@ export const vip525 = () => {
             ],
           ],
         ],
-        dstChainId: LzChainId.ethereum,
+        dstChainId: LzChainId.sepolia,
       },
     ],
     meta,
@@ -256,4 +255,4 @@ export const vip525 = () => {
   );
 };
 
-export default vip525;
+export default vip530;
